+++ src/main/java/chess/controller/ChessController.java
@@ -1,25 +1,18 @@
 package chess.controller;
 
-import chess.domain.command.Commands;
-import chess.domain.dto.MoveRequestDto;
-import chess.domain.dto.NameDto;
 import chess.domain.exception.DataException;
-import chess.domain.response.GameResponse;
 import chess.service.ChessService;
 import chess.view.ModelView;
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
-
-import java.sql.SQLException;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping("/play")
 public class ChessController {
 
-    private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
     private final ChessService chessService;
 
     public ChessController(ChessService chessService) {
@@ -30,12 +23,6 @@ public class ChessController {
     public String play(Model model) throws DataException {
         model.addAllAttributes(ModelView.historyResponse(chessService.loadHistory()));
         return "lobby";
-    }
-
-    @PostMapping("")
-    @ResponseBody
-    public String saveName(@RequestBody NameDto nameDto) {
-        return GSON.toJson(ModelView.idResponse(chessService.addHistory(nameDto.getName())));
     }
 
     @GetMapping("/{id}")
@@ -56,22 +43,5 @@ public class ChessController {
     @GetMapping("/end")
     public String endGame() {
         return "lobby";
-    }
-
-    @PostMapping("/move")
-    @ResponseBody
-    public String move(@RequestBody MoveRequestDto moveRequestDto) {
-        String command = makeMoveCmd(moveRequestDto.getSource(), moveRequestDto.getTarget());
-        String id = moveRequestDto.getGameId();
-        try {
-            chessService.move(id, command, new Commands(command));
-            return GSON.toJson(new GameResponse(chessService.continuedGameInfo(id), id));
-        } catch (IllegalArgumentException | SQLException e) {
-            return e.getMessage();
-        }
-    }
-
-    private String makeMoveCmd(String source, String target) {
-        return String.join(" ", "move", source, target);
     }
 }
